pyproject.toml
[project]
name = "multiagent-fitness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=1.2.20",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
